13-authentication-renew/pyproject.toml
[tool.poetry]
name = "13-authentication-renew"
version = "0.1.0"
description = ""
authors = ["Noah Kim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.19"
autopep8 = "^2.0.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
